split the nse_update functions out of aprox19 this will allow us to use them for both types of NSE we support
